// pokemon-albums/src/Store/types.ts
export enum ActionsTypes {
  GET_POKEMONS = 'GET_POKEMONS',

  TOGGLE_ENABLE_FORM_NEW_POKEMON = 'TOGGLE_ENABLE_FORM_NEW_POKEMON',

  SET_POKEMON_CREATED = 'SET_POKEMON_CREATED',
  RESET_POKEMON_CREATED = 'RESET_POKEMON_CREATED',

  SET_POKEMON_DELETED = 'SET_POKEMON_DELETED',
  RESET_POKEMON_DELETED = 'RESET_POKEMON_DELETED',

  TOGGLE_EDIT_MODE = 'TOGGLE_EDIT_MODE',

  SET_POKEMON_TO_EDIT = 'SET_POKEMON_TO_EDIT',
  SET_POKEMON_EDITED = 'SET_POKEMON_EDITED',
  RESET_POKEMON_EDITED = 'RESET_POKEMON_EDITED',
}